% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank}
\alias{bank}
\title{bank dataset}
\format{
\subsection{\code{bank}}{

A data frame with 100 rows and 3 columns:
\describe{
\item{...1}{customer number}
\item{arrival_time}{customer arrival time}
\item{service_time}{customer service time}
...
}
}
}
\source{
STAT1378
}
\usage{
bank
}
\description{
A subset of data from the bank
Report ...
}
\keyword{datasets}
